refactor: relocate telemetry configuration
